+++ tests/benchmarks/test_last_count.py
@@ -16,8 +16,8 @@ def _last(data: Range) -> int:
     return data.iter().last()
 
 
-def _length(data: Range) -> int:
-    return data.iter().length()
+def _count(data: Range) -> int:
+    return data.iter().count()
 
 
 @pytest.mark.benchmark(group="last")
@@ -27,8 +27,8 @@ def test_last(benchmark: BenchFixture, size: int) -> None:
     assert benchmark(_last, data) == size - 1
 
 
-@pytest.mark.benchmark(group="length")
+@pytest.mark.benchmark(group="count")
 @pytest.mark.parametrize("size", SIZES)
-def test_length(benchmark: BenchFixture, size: int) -> None:
+def test_count(benchmark: BenchFixture, size: int) -> None:
     data = Range(0, size)
-    assert benchmark(_length, data) == size
+    assert benchmark(_count, data) == size