+++ tests/test_last_count.py
@@ -16,11 +16,11 @@ def test_last_unique_by() -> None:
     assert data.iter().unique_by(len).last() == "mouse"
 
 
-def test_length_unique() -> None:
+def test_count_unique() -> None:
     data = Seq((1, 2, 1, 3))
-    assert data.iter().unique().length() == 3
+    assert data.iter().unique().count() == 3
 
 
-def test_length_unique_by() -> None:
+def test_count_unique_by() -> None:
     data = Seq(("cat", "mouse", "dog", "hen"))
-    assert data.iter().unique_by(len).length() == 2
+    assert data.iter().unique_by(len).count() == 2